+++ lsp/src/main/kotlin/com/falsepattern/zigbrains/lsp/settings/ZLSSettings.kt
@@ -31,6 +31,7 @@ data class ZLSSettings(
     @JvmField @Attribute val zlsConfigPath: @NonNls String = "",
     @JvmField @Attribute val inlayHints: Boolean = true,
     @JvmField @Attribute val inlayHintsMaxFileSizeKb: Int = 128,
+    @JvmField @Attribute val selectionRanges: Boolean = false,
     @JvmField @Attribute val enable_snippets: Boolean = true,
     @JvmField @Attribute val enable_argument_placeholders: Boolean = true,
     @JvmField @Attribute val completion_label_details: Boolean = true,